my-pets - Move DogCard into its own file so that it can become a client component
